// frontend/src/app/shared/enums/AchievementType.ts
export enum AchievementType {
    PROJECT = 'PROJECT',
    HOBBY = 'HOBBY',
    VOLUNTEER = 'VOLUNTEER',
    COMPETITION = 'COMPETITION',
    RECOGNITION = 'RECOGNITION',
    ACADEMIC = 'ACADEMIC',
    PROFESSIONAL = 'PROFESSIONAL',
    TECHNICAL = 'TECHNICAL',
    RESEARCH = 'RESEARCH',
    LEADERSHIP = 'LEADERSHIP',
    CERTIFICATION = 'CERTIFICATION',
    PUBLICATION = 'PUBLICATION',
    PATENT = 'PATENT',
    AWARD = 'AWARD'
}